+++ src/Builder/FormContractor.php
@@ -113,7 +113,7 @@ class FormContractor implements FormContractorInterface
                 ));
             }
 
-            $options['class'] = $fieldDescription->getTargetEntity();
+            $options['class'] = $fieldDescription->getTargetModel();
             $options['model_manager'] = $fieldDescription->getAdmin()->getModelManager();
 
             if ($this->checkFormClass($type, [ModelAutocompleteType::class])) {
@@ -122,7 +122,7 @@ class FormContractor implements FormContractorInterface
                         'The current field `%s` is not linked to an admin.'
                         .' Please create one for the target entity: `%s`',
                         $fieldDescription->getName(),
-                        $fieldDescription->getTargetEntity()
+                        $fieldDescription->getTargetModel()
                     ));
                 }
             }
@@ -132,7 +132,7 @@ class FormContractor implements FormContractorInterface
                     'The current field `%s` is not linked to an admin.'
                     .' Please create one for the target entity : `%s`',
                     $fieldDescription->getName(),
-                    $fieldDescription->getTargetEntity()
+                    $fieldDescription->getTargetModel()
                 ));
             }
 
@@ -165,7 +165,7 @@ class FormContractor implements FormContractorInterface
                     'The current field `%s` is not linked to an admin.'
                     .' Please create one for the target entity : `%s`',
                     $fieldDescription->getName(),
-                    $fieldDescription->getTargetEntity()
+                    $fieldDescription->getTargetModel()
                 ));
             }